requirements.txt
langchain==0.0.184
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
openai==0.27.6
faiss-cpu==1.7.4
altair<5
tiktoken==0.4.0
numpy==1.24.3
image==1.5.33
streamlit-extras==0.2.7
requests==2.31.0